requirements_v3.txt
Flask==3.0.0
flask-cors==4.0.0
python-docx==1.1.0
python-pptx==0.6.23
pymupdf
pywin32>=306
reportlab==4.0.7
markdown==3.5.1
Werkzeug==3.0.1
google-generativeai>=0.3.0